♻️ bot/_legacy 에서 scheduling 모듈 분리 (P0-Q step 5)
변경 이유:
- bot/_legacy.py 3126줄 중 일정 스케줄링/체크포인트/브리핑 데이터 cache 헬퍼는 응집도가 높아 한 덩어리로 묶을 수 있음
- 후속 message_routing / engineering_handlers / main 분해 시 scheduling 만 import 하도록 의존성 축소

주요 변경 사항:
- src/yule_orchestrator/discord/bot/scheduling.py 신설 (340줄, 21개 함수 + 7개 상수)
  · 일별 준비 tick: _next_daily_run, _collect_due_daily_preparation_steps, _daily_preparation_schedule_for, _cleanup_completed_preparation_steps, _next_daily_preparation_runs
  · 체크포인트 스캔: _next_checkpoint_scan, _checkpoint_channel_error_label, _checkpoint_window_minutes, _resolve_due_checkpoints, _filter_unsent_checkpoints, _mark_checkpoints_sent, _has_checkpoint_been_sent, _checkpoint_cache_key
  · 정기 브리핑 윈도우 + cache: _next_scheduled_briefing_run, _resolve_due_briefings, _collect_due_briefing_slots, _synthesize_scheduled_briefing, _has_briefing_been_sent_async, _briefing_notification_cache_key, _filter_unsent_briefings, _mark_briefings_sent
  · discord-send metric: _save_discord_send_metric
- _legacy.py 에서 동일 정의 삭제 + scheduling 에서 re-import (facade attribute 유지)
- test_bot_runtime.py: build_due_checkpoints / load_prefetched_due_checkpoints patch 경로를 bot.scheduling.* 로 (call site 이동)

비고:
- 4672 PASS / 1 skipped — 회귀 0건
- 외부 import 경로 (bot 패키지 facade) 그대로
- DAILY_PREPARATION_GITHUB_LIMIT 은 _legacy.py 내부에서만 쓰여 거기에 잔류
